requirements.txt
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.0
openai==1.3.5
pandas==2.0.3
flask==2.3.3
tiktoken==0.5.1
pymongo
openai
tensorflow
numpy
Pillow
Flask
scipy
streamlit
gpt3-tokenizer
fontawesome
